he reecho el control de errores de los argumentos y simplificado la forma de controlar que los argumentos sean correctos

+++ libft/src/ft_str/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ciclo <ciclo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/12 01:26:33 by ciclo             #+#    #+#             */
-/*   Updated: 2022/10/24 12:59:30 by ciclo            ###   ########.fr       */
+/*   Updated: 2022/11/16 19:54:53 by ciclo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 /// @return  The length of the string s.
 size_t	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != 0)
